pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-assistant-deepagents"
version = "0.32.0"
description = "Autonomous code review agent powered by LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "chronicqazxc@gmail.com"}
]
dependencies = [
    "deepagents>=0.0.10",
    "langchain-ollama",
    "langchain-anthropic",
    "langchain-google-genai",
    "langchain-community",
    "anthropic>=0.42.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "json-repair>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
pr-assistant = "pr_assistant.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pr_assistant*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
